+++ src/main/java/io/pet/cocotrip/BeApplication.java
@@ -1,0 +1,15 @@
+package io.pet.cocotrip;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@EnableJpaAuditing
+@SpringBootApplication
+public class BeApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(BeApplication.class, args);
+    }
+
+}
