+++ src/main/java/io/pet/cocotrip/domain/auth/application/dto/CreateAccessTokenRequest.java
@@ -1,0 +1,10 @@
+package io.pet.cocotrip.domain.auth.application.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class CreateAccessTokenRequest {
+    private String refreshToken;
+}
