+++ src/main/java/io/pet/cocotrip/domain/auth/application/dto/CreateAccessTokenResponse.java
@@ -1,0 +1,10 @@
+package io.pet.cocotrip.domain.auth.application.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class CreateAccessTokenResponse {
+    private String accessToken;
+}
