+++ src/main/java/io/pet/cocotrip/domain/auth/persistence/RefreshTokenRepository.java
@@ -1,0 +1,10 @@
+package io.pet.cocotrip.domain.auth.persistence;
+
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+    Optional<RefreshToken> findByUserId(Long userId);
+
+    Optional<RefreshToken> findByRefreshToken(String refreshToken);
+}
