+++ src/main/java/io/pet/cocotrip/domain/user/application/dto/AddUserRequest.java
@@ -1,0 +1,11 @@
+package io.pet.cocotrip.domain.user.application.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class AddUserRequest {
+    private String email;
+    private String password;
+}
