+++ src/main/java/io/pet/cocotrip/global/config/filter/TokenAuthenticationFilter.java
@@ -1,0 +1,63 @@
+package io.pet.cocotrip.global.config.filter;
+
+import io.pet.cocotrip.global.config.jwt.TokenProvider;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+/**
+ * TokenAuthenticationFilter 토큰 인증 필터
+ */
+@RequiredArgsConstructor
+public class TokenAuthenticationFilter extends OncePerRequestFilter {
+    private final TokenProvider tokenProvider;
+    private final static String HEADER_AUTHORIZATION = "Authorization";
+    private final static String TOKEN_PREFIX = "Bearer ";
+
+    /**
+     * 토큰 인증 필터
+     *
+     * @param request     HttpServletRequest 객체
+     * @param response    HttpServletResponse 객체
+     * @param filterChain FilterChain 객체
+     * @throws ServletException 서블릿 예외
+     * @throws IOException      IO 예외
+     */
+    @Override
+    protected void doFilterInternal(
+            HttpServletRequest request,
+            HttpServletResponse response,
+            FilterChain filterChain) throws ServletException, IOException {
+
+        //요청 헤더의 Authorization 키의 값 조회
+        String authorizationHeader = request.getHeader(HEADER_AUTHORIZATION);
+        // 가져온 값에서 접두사 제거
+        String token = getAccessToken(authorizationHeader);
+        // 가져온 토큰이 유효한지 확인하고, 유효한 때는 인증 정보 설정
+        if (tokenProvider.validToken(token)) {
+            Authentication authentication = tokenProvider.getAuthentication(token);
+            SecurityContextHolder.getContext().setAuthentication(authentication);
+        }
+
+        filterChain.doFilter(request, response);
+    }
+
+    /**
+     * 헤더에서 Bearer 을 제거하고 토큰 추출
+     *
+     * @param authorizationHeader Authorization 헤더 값
+     * @return 토큰
+     */
+    private String getAccessToken(final String authorizationHeader) {
+        if (authorizationHeader != null && authorizationHeader.startsWith(TOKEN_PREFIX)) {
+            return authorizationHeader.substring(TOKEN_PREFIX.length());
+        }
+        return null;
+    }
+}
