--- src/main/java/io/pet/travel/domain/auth/application/dto/CreateAccessTokenRequest.java
@@ -1,10 +1,0 @@
-package io.pet.travel.domain.auth.application.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class CreateAccessTokenRequest {
-    private String refreshToken;
-}
