--- src/main/java/io/pet/travel/domain/auth/application/dto/CreateAccessTokenResponse.java
@@ -1,10 +1,0 @@
-package io.pet.travel.domain.auth.application.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class CreateAccessTokenResponse {
-    private String accessToken;
-}
