--- src/main/java/io/pet/travel/domain/auth/persistence/entity/RefreshToken.java
@@ -1,36 +1,0 @@
-package io.pet.travel.domain.auth.persistence.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Entity
-public class RefreshToken {
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    @Column(name = "id", updatable = false)
-    private Long id;
-
-    @Column(name = "user_id", nullable = false, unique = true)
-    private Long userId;
-
-    @Column(name = "refresh_token", nullable = false)
-    private String refreshToken;
-
-    public RefreshToken(final Long userId, final String refreshToken) {
-        this.userId = userId;
-        this.refreshToken = refreshToken;
-    }
-
-    public RefreshToken update(final String newRefreshToken) {
-        this.refreshToken = newRefreshToken;
-        return this;
-    }
-}
