--- src/main/java/io/pet/travel/domain/auth/persistence/repository/RefreshTokenRepository.java
@@ -1,11 +1,0 @@
-package io.pet.travel.domain.auth.persistence.repository;
-
-import io.pet.travel.domain.auth.persistence.entity.RefreshToken;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
-    Optional<RefreshToken> findByUserId(Long userId);
-
-    Optional<RefreshToken> findByRefreshToken(String refreshToken);
-}
