--- src/main/java/io/pet/travel/domain/user/dto/AddUserRequest.java
@@ -1,11 +1,0 @@
-package io.pet.travel.domain.user.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class AddUserRequest {
-    private String email;
-    private String password;
-}
