+++ src/test/java/io/pet/cocotrip/BeApplicationTests.java
@@ -1,0 +1,13 @@
+package io.pet.cocotrip;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BeApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
