+++ src/test/java/io/pet/cocotrip/config/jwt/TokenProviderTest.java
@@ -1,0 +1,112 @@
+package io.pet.cocotrip.config.jwt;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import io.jsonwebtoken.Jwts;
+import io.pet.cocotrip.domain.user.persistence.User;
+import io.pet.cocotrip.domain.user.persistence.UserRepository;
+import io.pet.cocotrip.global.config.jwt.JwtProperties;
+import io.pet.cocotrip.global.config.jwt.TokenProvider;
+import java.time.Duration;
+import java.util.Date;
+import java.util.Map;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.userdetails.UserDetails;
+
+@SpringBootTest
+public class TokenProviderTest {
+    @Autowired
+    private TokenProvider tokenProvider;
+    @Autowired
+    private UserRepository userRepository;
+    @Autowired
+    private JwtProperties jwtProperties;
+
+    @DisplayName("generateToken() : 유저 정보와 만료 기간을 전달해 토큰을 만들 수 있다.")
+    @Test
+    void generateToken() {
+        // given
+        User testUser = userRepository.save(User.builder()
+                .email("user@gmail.com")
+                .password("test")
+                .build());
+
+        // when
+        String token = tokenProvider.generateToken(testUser, Duration.ofDays(14));
+
+        //then
+        Long userId = Jwts.parser()
+                .setSigningKey(jwtProperties.getSecretKey())
+                .parseClaimsJws(token)
+                .getBody()
+                .get("id", Long.class);
+        assertThat(userId).isEqualTo(testUser.getUserId());
+    }
+
+    @DisplayName("validToken() : 만료된 토큰인 때에 유효성 검증에 실패한다.")
+    @Test
+    void validToken_invalidToken() {
+        //given
+        String token = JwtFactory.builder()
+                .expiration(new Date(new Date().getTime() - Duration.ofDays(7).toMillis()))
+                .build()
+                .createToken(jwtProperties);
+
+        //when
+        boolean result = tokenProvider.validToken(token);
+
+        //then
+        assertThat(result).isFalse();
+    }
+
+    @DisplayName("validToken() : 유효한 토큰일 때에 유효성 검증에 성공한다.")
+    @Test
+    void validToken_validToken() {
+        //given
+        String token = JwtFactory.withDefaultValues().createToken(jwtProperties);
+
+        //when
+        boolean result = tokenProvider.validToken(token);
+
+        //then
+        assertThat(result).isTrue();
+    }
+
+    @DisplayName("getAuthentication() : 토큰 기반으로 인증 정보를 가져올 수 있다.")
+    @Test
+    void getAuthentication() {
+        //given
+        String userEmail = "user@email.com";
+        String token = JwtFactory.builder()
+                .subject(userEmail)
+                .build()
+                .createToken(jwtProperties);
+
+        //when
+        Authentication authentication = tokenProvider.getAuthentication(token);
+
+        //then
+        assertThat(((UserDetails) authentication.getPrincipal()).getUsername()).isEqualTo(userEmail);
+    }
+
+    @DisplayName("getUserId() : 토큰으로 유저 ID를 가져올 수 있다.")
+    @Test
+    void getUserId() {
+        //given
+        Long userId = 1L;
+        String token = JwtFactory.builder()
+                .claims(Map.of("id", userId))
+                .build()
+                .createToken(jwtProperties);
+
+        //when
+        Long result = tokenProvider.getUserId(token);
+
+        //then
+        assertThat(result).isEqualTo(userId);
+    }
+}
