--- src/test/java/io/pet/travel/BeApplicationTests.java
@@ -1,13 +1,0 @@
-package io.pet.travel;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BeApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
