--- src/test/java/io/pet/travel/controller/TokenApiControllerTest.java
@@ -1,94 +1,0 @@
-package io.pet.travel.controller;
-
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import io.pet.travel.config.jwt.JwtFactory;
-import io.pet.travel.domain.auth.application.dto.CreateAccessTokenRequest;
-import io.pet.travel.domain.auth.persistence.entity.RefreshToken;
-import io.pet.travel.domain.auth.persistence.repository.RefreshTokenRepository;
-import io.pet.travel.domain.user.entity.User;
-import io.pet.travel.domain.user.repository.UserRepository;
-import io.pet.travel.global.config.jwt.JwtProperties;
-import java.util.Map;
-import io.pet.travel.config.jwt.JwtFactory;
-import io.pet.travel.domain.auth.application.dto.CreateAccessTokenRequest;
-import io.pet.travel.domain.auth.persistence.entity.RefreshToken;
-import io.pet.travel.domain.auth.persistence.repository.RefreshTokenRepository;
-import io.pet.travel.domain.auth.web.controller.TokenApiController;
-import io.pet.travel.domain.user.entity.User;
-import io.pet.travel.domain.user.repository.UserRepository;
-import io.pet.travel.global.config.jwt.JwtProperties;
-import java.util.Map;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.http.MediaType;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.ResultActions;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.context.WebApplicationContext;
-
-//@WebMvcTest(TokenApiController.class) // 작동안됨
-@SpringBootTest
-@AutoConfigureMockMvc
-public class TokenApiControllerTest {
-    @Autowired
-    protected MockMvc mockMvc;
-    @Autowired
-    protected ObjectMapper objectMapper;
-    @Autowired
-    private WebApplicationContext context;
-    @Autowired
-    JwtProperties jwtProperties;
-    @Autowired
-    UserRepository userRepository;
-    @Autowired
-    RefreshTokenRepository refreshTokenRepository;
-
-    @BeforeEach
-    public void mockMvcSetup() {
-        this.mockMvc = MockMvcBuilders.webAppContextSetup(context)
-                .build();
-        userRepository.deleteAll();
-    }
-
-    @DisplayName("createNewAccessToken : 새로운 엑세스 토큰을 생성한다.")
-    @Test
-    public void createNewAccessToken() throws Exception {
-        // given
-        final String url = "/api/token";
-
-        User testUser = userRepository.save(User.builder()
-                .email("user@gmail.com")
-                .password("test")
-                .build());
-
-        String refreshToken = JwtFactory.builder()
-                .claims(Map.of("id", testUser.getUserId()))
-                .build()
-                .createToken(jwtProperties);
-
-        refreshTokenRepository.save(new RefreshToken(testUser.getUserId(), refreshToken));
-
-        CreateAccessTokenRequest request = new CreateAccessTokenRequest();
-        request.setRefreshToken(refreshToken);
-        final String requestBody = objectMapper.writeValueAsString(request);
-
-        // when
-        ResultActions resultActions = mockMvc.perform(post(url)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .content(requestBody));
-        // then
-        resultActions
-                .andExpect(status().isCreated())
-                .andExpect(jsonPath("$.accessToken").exists());
-    }
-}
-
