--- src/test/java/io/pet/travel/domain/user/repository/UserRepositoryTest.java
@@ -1,37 +1,0 @@
-package io.pet.travel.domain.user.repository;
-
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-
-import io.pet.travel.domain.user.entity.User;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.annotation.Rollback;
-import org.springframework.transaction.annotation.Transactional;
-
-@SpringBootTest
-@Transactional
-@Rollback(value = false)
-public class UserRepositoryTest {
-
-    @Autowired
-    UserRepository userRepository;
-
-    @Test
-    @DisplayName("Auditing 기능적용 테스트")
-    public void findUser() {
-        //given
-        User user = User.builder()
-                .nickname("test")
-                .email("test@example.com")
-                .build();
-
-        //when
-        User savedUser = userRepository.save(user);
-
-        //then
-        assertNotNull(savedUser.getCreatedAt());
-        assertNotNull(savedUser.getUpdatedAt());
-    }
-}
